Enum gdt::Descriptor

pub enum Descriptor {
    UserSegment(u64),
    SystemSegment(u64, u64),
}

The two kinds of descriptor entries in the GDT.

Variants

UserSegment(u64)

UserSegment is used for both code and data segments, in both the kernel and in user space.

SystemSegment(u64, u64)

SystemSegment is used only for TSS.

Implementations

impl Descriptor

pub const fn kernel_code_segment() -> Descriptor

pub const fn kernel_data_segment() -> Descriptor

pub const fn user_code_32_segment() -> Descriptor

pub const fn user_data_32_segment() -> Descriptor

pub const fn user_code_64_segment() -> Descriptor

pub const fn user_data_64_segment() -> Descriptor

pub fn tss_segment(tss: &TaskStateSegment) -> Descriptor